+++ IDGS903_Api/Models/materiaPrima.cs
@@ -7,8 +7,8 @@ namespace IDGS903_Api.Models
         [Key]
         public int Id { get; set; }
         public string? Nombre { get; set; }
-        public float Costo { get; set; }
-        public float Stock { get; set; }
+        public Double Costo { get; set; }
+        public Double Stock { get; set; }
         public int TipoMateriaPrima_id { get; set; }
         public int Proveedor_id { get; set; }
